# internal/server/openapi.go
package server

import (
	"net/http"

	httpSwagger "github.com/swaggo/http-swagger/v2"

	_ "git.erwanleboucher.dev/eleboucher/apoci/internal/apidocs" // generated by `make docs`
	"git.erwanleboucher.dev/eleboucher/apoci/internal/peering"
)

// GCStatusResponse lets swag resolve the /gc schema; admin.go cannot import peering.
type GCStatusResponse = peering.GCStatus

// @title        apoci admin API
// @version      1.0
// @description  Administrative API for an apoci federated OCI registry node.
// @BasePath     /api/admin
//
// @securityDefinitions.apikey  Bearer
// @in                          header
// @name                        Authorization
// @description                 Admin token as "Bearer <token>".

func swaggerHandler() http.Handler {
	return httpSwagger.Handler(httpSwagger.URL("/api/admin/docs/doc.json"))
}